+++ packages/terra-icon/src/IconBase.jsx
@@ -4,7 +4,7 @@ import classNames from 'classnames/bind';
 import 'terra-base/lib/baseStyles';
 
 // eslint-disable-next-line import/no-unresolved, import/no-webpack-loader-syntax
-import styles from './Icon.scss';
+import styles from './Icon.module.scss';
 
 const cx = classNames.bind(styles);